worker_sim/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.42.0
google-generativeai==0.5.2
python-dotenv==1.0.1
lxml==5.2.2
